update delete pictures book et user

+++ src/Controller/Admin/AdminController.php
@@ -26,11 +26,17 @@ class AdminController extends AbstractController
 
         ]);
     }
-    #[Route('/{id}', name: 'user_delete', methods: ['POST', 'GET'])]
-    public function delete(User $user, UserRepository $userRepository): Response
+    #[Route('/delete/{id}', name: 'user_delete', methods: ['POST', 'GET'])]
+    public function delete(int $id, UserRepository $userRepository): Response
     {
+        $user = $userRepository->find($id);
+
+        if (!$user) {
+            throw $this->createNotFoundException('Utilisateur introuvable');
+        }
+
         $userRepository->remove($user, true);
-        $this->addFlash('success', 'user supprimé');
+        $this->addFlash('success', 'Utilisateur supprimé');
         return $this->redirectToRoute('app_admin', [], Response::HTTP_SEE_OTHER);
     }
 }
